+++ sdk/appconfiguration/azure-appconfiguration/azure/appconfiguration/aio/_feature_flag_client_async.py
@@ -396,10 +396,9 @@ class FeatureFlagClient:
         after: Optional[str] = None,
         accept_datetime: Optional[Union[datetime, str]] = None,
         fields: Optional[List[Union[str, LabelFields]]] = None,
-        resource_type: Optional[str] = None,
         **kwargs: Any,
     ) -> AsyncItemPaged[ConfigurationSettingLabel]:
-        """Gets a list of labels.
+        """Gets a list of feature flag labels.
 
         :keyword name: A filter for the name of the returned labels.  '*' can be used as wildcard
             in the beginning or end of the filter. For more information about supported filters, see
@@ -414,9 +413,6 @@ class FeatureFlagClient:
         :keyword fields: Specify which fields to include in the results. If not specified, will include all fields.
             Available fields see :class:`~azure.appconfiguration.LabelFields`.
         :paramtype fields: list[str] or list[~azure.appconfiguration.LabelFields] or None
-        :keyword resource_type: A filter used to indicate the resource type of the labels. Accepted values:
-            'kv' for key-value labels, 'ff' for feature flag labels. Default is `None`.
-        :paramtype resource_type: str or None
         :return: An async iterator of labels.
         :rtype: ~azure.core.paging.AsyncItemPaged[~azure.appconfiguration.ConfigurationSettingLabel]
         :raises: :class:`~azure.core.exceptions.HttpResponseError`
@@ -428,7 +424,7 @@ class FeatureFlagClient:
             after=after,
             accept_datetime=accept_datetime,
             select=fields,
-            resource_type=resource_type,
+            resource_type="ff",
             cls=lambda objs: [ConfigurationSettingLabel(name=x.name) for x in objs],
             **kwargs,
         )